refactor global and company services
