+++ front-angular/src/app/logging/logging.component.ts
@@ -4,7 +4,7 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { first } from 'rxjs/operators';
-import { GlobalService } from '../global.service';
+import { GlobalService } from '../services/global.service';
 
 
 @Component({
