+++ front-angular/src/app/navbar/navbar.component.ts
@@ -2,7 +2,7 @@ import { NotificationsService } from './../profile/notification/notifications.se
 import { Component, OnInit } from '@angular/core';
 import { User } from 'src/models';
 import { AuthenticationService } from '../logging/services';
-import { GlobalService } from '../global.service';
+import { GlobalService } from '../services/global.service';
 import { Subscription } from 'rxjs';
 
 @Component({
@@ -14,9 +14,9 @@ export class NavbarComponent implements OnInit {
   currentUser: User;
   isProfilOpen: Boolean;
   isProfilOpenSubscription: Subscription;
-  nbrNotif:number;
-  nbrNotifSubscription :Subscription;
-  constructor(private authenticationService: AuthenticationService, private globalService: GlobalService, private notificationsService : NotificationsService
+  nbrNotif: number;
+  nbrNotifSubscription: Subscription;
+  constructor(private authenticationService: AuthenticationService, private globalService: GlobalService, private notificationsService: NotificationsService
   ) {
     this.authenticationService.currentUser.subscribe(x => this.currentUser = x);
   }
