+++ front-angular/src/app/profile/application/offer-company/add-offer/add-offer.component.ts
@@ -14,7 +14,7 @@ import { MatDatepicker } from '@angular/material';
 import { SelectOption } from 'src/models/SelectOption';
 import { AuthenticationService } from 'src/app/logging/services';
 import { User } from 'src/models/user';
-import { CompanyService } from 'src/app/company.service';
+import { CompanyService } from '../../../../services/company.service';
 import { Company } from 'src/models';
 const moment = _rollupMoment || _moment;
 
@@ -74,9 +74,9 @@ export class AddOfferComponent implements OnInit {
 
   currentUser: any;
   constructor(private offerViewService: OfferViewService,
-              private authenticationService: AuthenticationService,
-              private companyService: CompanyService,
-              private router : Router) {
+    private authenticationService: AuthenticationService,
+    private companyService: CompanyService,
+    private router: Router) {
     this.authenticationService.currentUser.subscribe(x => this.currentUser = x);
   }
 
@@ -116,10 +116,10 @@ export class AddOfferComponent implements OnInit {
     });
     this.companyService.getAll().subscribe(
       value => {
-          this.companiesList = value;
+        this.companiesList = value;
       },
       error => {
-          console.log('Erreur ! : ' + error);
+        console.log('Erreur ! : ' + error);
       }
     );
   }
@@ -141,16 +141,16 @@ export class AddOfferComponent implements OnInit {
       this.offerOnForm.start_date = '' + this.dateFromDate.getTime();
       this.offerOnForm.created_date = '' + (new Date()).getTime(); //TODO : Changer les types pour que rien soit cassé même si ça fonctionne
       this.offerOnForm.location = this.locationCity + ', ' + this.locationCountry;
-      if (!this.offerOnForm.isValid()){
-        this.errorOnForm=true;
+      if (!this.offerOnForm.isValid()) {
+        this.errorOnForm = true;
         return;
       }
       this.offerViewService.addOffer(this.offerOnForm);
     } else {
       this.offerOnForm.start_date = '' + this.dateFromDate.getTime();
       this.offerOnForm.location = this.locationCity + ', ' + this.locationCountry;
-      if (!this.offerOnForm.isValid()){
-        this.errorOnForm=true;
+      if (!this.offerOnForm.isValid()) {
+        this.errorOnForm = true;
         return;
       }
       this.offerViewService.editOffer(this.offerOnForm);
