+++ front-angular/src/app/profile/profile.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { AuthenticationService } from '../logging/services';
-import { User } from 'src/models';
 import { Router } from '@angular/router';
-import { GlobalService } from '../global.service';
+import { GlobalService } from '../services/global.service';
 
 @Component({
   selector: 'app-profile',
