+++ front-angular/src/app/services/company.service.ts
@@ -1,14 +1,12 @@
-import { Injectable } from "@angular/core";
+import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Subject } from 'rxjs/Subject';
-import { Company } from '../models/Company';
+import { Company } from '../../models/Company';
 import { environment } from 'src/environments/environment';
 
 @Injectable()
 export class CompanyService {
 
-
-    constructor(private http: HttpClient) {}
+    constructor(private http: HttpClient) { }
 
     apiUrl = environment.apiUrl;
 
@@ -17,7 +15,6 @@ export class CompanyService {
     }
 
     getById(id: any) {
-        return this.http.get<Company>(this.apiUrl + '/companies/'+id)
+        return this.http.get<Company>(this.apiUrl + '/companies/' + id);
     }
-
 }
